feat: refactor and add a single database reader using sqlalchemy for mysql, postgresql, sqlite
